ADC battery read fix
`BATTREAD` terminal command now prints battery voltage reading using ADC in float format. Tested with a test bench power supply of 3.8V and sensor outputs 3.78

+++ FloodnetDriver/Adc.h
@@ -10,7 +10,7 @@
 #include "main.h"
 
 
-uint16_t AdcRead_VBatt(void);
+float AdcRead_VBatt(void);
